OFT/BOFT support, wd on output support, unified weight adapter solution (#357)
